Directory and files base structure rebuild (#354)
