+++ yajsapi/agreement/index.ts
@@ -1,0 +1,3 @@
+import { AgreementPoolService } from "./agreement_pool_service";
+
+export { AgreementPoolService };
