+++ yajsapi/market/market_service.ts
@@ -1,0 +1,1 @@
+export class MarketService {}
