+++ yajsapi/market/strategy.ts
@@ -1,0 +1,3 @@
+export interface MarketStrategy {
+  todo: true;
+}
