+++ yajsapi/payment/debit_note.ts
@@ -1,0 +1,1 @@
+export class DebitNote {}
