+++ yajsapi/payment/invoice.ts
@@ -1,0 +1,1 @@
+export class Invoice {}
