--- yajsapi/script/index.ts
@@ -1,3 +1,0 @@
-import { Script } from "./script";
-import { Command, Run, Deploy, Start, Transfer, Terminate, UploadFile, DownloadFile, Capture } from "./command";
-export { Script, Command, Run, Deploy, Start, Transfer, Terminate, UploadFile, DownloadFile, Capture };
