--- yajsapi/utils/callable.ts
@@ -1,3 +1,0 @@
-export default interface Callable<T, R> {
-  (...T): R;
-}
